requirements.txt
Flask==3.0.0
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.2
pdfplumber==0.10.3
python-docx==1.1.0
easyocr==1.7.2
Pillow==12.0.0
pyodbc==5.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
xlrd==2.0.1
openai>=1.0.0
httpx>=0.27.0
python-dotenv==1.0.0

